REPO-3616: Move heartbeat job classes to a new jobs package
