+++ src/main/java/org/alfresco/heartbeat/jobs/LockingJobScheduler.java
@@ -23,8 +23,9 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.heartbeat;
+package org.alfresco.heartbeat.jobs;
 
+import org.alfresco.heartbeat.HBBaseDataCollector;
 import org.alfresco.repo.lock.JobLockService;
 import org.quartz.Job;
 import org.quartz.JobDataMap;
