+++ src/main/resources/alfresco/heartbeat/heartbeat-context.xml
@@ -28,7 +28,7 @@
 
     <!-- The default HeartBeat job scheduler, this scheduler schedules jobs which lock execution on collector id.
     Only one repository node in a cluster will send data for collectors with this type of job -->
-    <bean id="hbLockingJobScheduler" class="org.alfresco.heartbeat.LockingJobScheduler">
+    <bean id="hbLockingJobScheduler" class="org.alfresco.heartbeat.jobs.LockingJobScheduler">
         <property name="testMode" value="${heartbeat.collectors.testMode}"/>
         <property name="scheduler" ref="schedulerFactory" />
         <property name="hbDataSenderService" ref="hbDataSenderService" />
@@ -37,7 +37,7 @@
 
     <!-- This scheduler schedules jobs which execute without any locking.
     The scheduler can be added to a collector by setting the hbJobScheduler property -->
-    <bean id="hbNonLockingJobScheduler" class="org.alfresco.heartbeat.NonLockingJobScheduler">
+    <bean id="hbNonLockingJobScheduler" class="org.alfresco.heartbeat.jobs.NonLockingJobScheduler">
         <property name="testMode" value="${heartbeat.collectors.testMode}"/>
         <property name="scheduler" ref="schedulerFactory" />
         <property name="hbDataSenderService" ref="hbDataSenderService" />
