+++ src/test/java/org/alfresco/heartbeat/jobs/QuartzJobSchedulerTest.java
@@ -23,9 +23,11 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.heartbeat;
+package org.alfresco.heartbeat.jobs;
 
+import org.alfresco.heartbeat.HBBaseDataCollector;
 import org.alfresco.heartbeat.datasender.HBData;
+import org.alfresco.heartbeat.jobs.QuartzJobScheduler;
 import org.alfresco.repo.scheduler.AlfrescoSchedulerFactory;
 import org.junit.Before;
 import org.junit.Test;
